feat: use SIMD dot product for binary and quaternary
